[ctags] Improve PostgreSQL SQL dialect support.
Using --langmap=sql was a mistake, because Ctags has a predefined sql
type, and the two were mixed, with side effects (e.g., duplicated tags
for functions).

This change requires an update of Vim configuration. In particularo,
for Tagbar, you need to put something like this in your .vimrc:

    let g:tagbar_type_sql = {
          \ 'ctagstype': 'pgsql',
          \ 'kinds': [
          \ 'f:functions',
          \ 'T:triggers',
          \ 't:tables',
          \ 'V:views',
          \ 'i:indexes',
          \ 'r:rules',
          \ 'd:types',
          \ 'S:sequences',
          \ 's:schemas',
          \ 'D:databases'
          \ ],
          \ 'sort': 0
          \ }

For CtrlPButTag, set:

    let g:ctrlp_buftag_types['sql'] = '--language-force=pgsql'